+++ catebot/__init__.py
@@ -1,0 +1,5 @@
+from .catebot import Catebot
+
+
+async def setup(bot):
+    await bot.add_cog(Catebot(bot))