+++ utils/createDatabase.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-sqlite3 status.db < status.sql